replace axios with native fetch

+++ eventHandlers/fileUploader.js
@@ -4,9 +4,8 @@
 */
 
 
-const axios = require('axios');
-const {GoogleAIFileManager} = require("@google/generative-ai/server");
-const {loadConfig} = require('../initializers/configuration');
+const { GoogleAIFileManager } = require("@google/generative-ai/server");
+const { loadConfig } = require('../initializers/configuration');
 const config = loadConfig();
 const fileManager = new GoogleAIFileManager(config.GEMINI_API_KEY);
 const log = require('../utils/betterLogs');
@@ -26,16 +25,17 @@ async function uploadFilesToGemini(message, client) {
                 const fPath = path.join(global.dirname, 'data', 'running', 'tmp', attachment.name);
                 try {
                     await message.react(state.emojis['uploading']);
-                    const response = await axios.get(attachment.url, {
-                        responseType: 'arraybuffer',
-                        validateStatus: () => true
-                    });
-                    fs.writeFileSync(fPath, response.data);
-                    const uploadedFile = await uploadToGemini(fPath, attachment.contentType);
-                    fs.unlinkSync(fPath);
-                    await message.reactions.cache.find(reaction => reaction.emoji.id === state.emojis['uploading'].id)?.users.remove(client.user.id);
-                    await message.react(state.emojis['uploaded']);
-                    return uploadedFile;
+                    const response = await fetch(attachment.url);
+                    if (response.ok) {
+                        fs.writeFileSync(fPath, response.data);
+                        const uploadedFile = await uploadToGemini(fPath, attachment.contentType);
+                        fs.unlinkSync(fPath);
+                        await message.reactions.cache.find(reaction => reaction.emoji.id === state.emojis['uploading'].id)?.users.remove(client.user.id);
+                        await message.react(state.emojis['uploaded']);
+                        return uploadedFile;
+                    } else {
+                        return null;
+                    }
                 } catch (error) {
                     console.error(`Error uploading file: ${error}`);
                     return null;
